Rename the notes use_cases package to commands
With both note reads in queries/, use_cases/ holds only create, update and
delete, so it takes the commands/ name per the ADR's end-state naming. The
dead-ends contract already covers both package names, so no contract edit is
needed.

parse_note_kind stays with the commands and is imported by the per-book read:
turning the API's kind string into a NoteKind is the same job on both sides, and
queries importing from commands is the allowed direction of that contract.
