+++ backend/src/application/notes/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Notes commands (write side)."""
