+++ backend/src/application/notes/commands/delete_note_use_case.py
