+++ backend/src/application/notes/commands/update_note_use_case.py
@@ -3,8 +3,8 @@
 import structlog
 
 from src.application.library.protocols.chapter_repository import ChapterRepositoryProtocol
+from src.application.notes.commands.helpers import parse_note_kind, validate_link_targets
 from src.application.notes.protocols.note_repository import NoteRepositoryProtocol
-from src.application.notes.use_cases.helpers import parse_note_kind, validate_link_targets
 from src.application.reading.protocols.highlight_repository import HighlightRepositoryProtocol
 from src.application.tagging.protocols.tag_repository import (
     TagRepositoryProtocol,
