--- backend/src/application/notes/use_cases/__init__.py
@@ -1,1 +1,0 @@
-"""Notes use cases."""
